syndication/requirements.txt
python-twitter
python-frontmatter
Jinja2
PyGithub
